Refactor: Calculator 코드 리팩토링
1. 함수명 변경
2. 라인 들여쓰기 및 주석 추가
