+++ README.md
@@ -20,7 +20,7 @@
 ## 예외
 - [x] 숫자 이외의 값, 음수 전달 시 RuntimeException throw
 
-=======
+============
 ----
 ## 기능 요구 사항
