+++ src/test/java/calculator/CustomSeparatorSumTest.java
@@ -11,6 +11,6 @@ public class CustomSeparatorSumTest {
     @DisplayName("커스텀 구분자를 기준으로 숫자의 합을 반환 ('//' '\n')")
     void customSum() {
         assertEquals(3, Calculator.sum("//;\n1;2"));
-        assertEquals(6, Calculator.sum("//!@\n1!2@3"));
+        assertEquals(6, Calculator.sum("//!\n1!2!3"));
     }
 }
