+++ src/test/java/car/racingtest/CarTest.java
@@ -1,4 +1,4 @@
-package car.logictest;
+package car.racingtest;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.api.Assertions.assertThrows;
