+++ src/test/java/car/racingtest/CarsTest.java
@@ -1,4 +1,4 @@
-package car.logictest;
+package car.racingtest;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertThrows;
